Modified Files: Tag: glite-data-hydra-cli_B_3_2_1_9 src/c/Makefile.am
This line fixes a not-yet-understood problem in that the
Authorization exceptions are not properly decoded.
This line removes two lines of autogenerated code in metadataClient.c
that do not behave properly.
It's possible this fails in the future as "_param_1" may
become eg. "_param_2" and therefore will be back to the same situation.
This needs to be debugged properly in the future.

Need to make this fix in order to release!